fix(titan): Process `waistBalance` similarly to `legBalance`
[The docs](https://github.com/freesewing/freesewing/blob/d2079515de02719501e9f3ff35cc7d7741c92b56/markdown/org/docs/patterns/titan/options/waistbalance/en.md#L11) suggest that the `waistBalance` option merely shifts the side seam at the waist level, but Titan's current code (1) does not alter the distribution of waist width between front and back, and (2) instead skews all the fabric above the crotch level, even close to CF/CB. This commit introduces more intuitive logic for the `waistBalance` option. Now, the vast majority of the garment will be unaffected, with the only changes occuring along the side seam.

Note that this commit uses a common basis (`measurements.knee / 2`) for the front and back waist-width adjustments (instead of `seatFront - waistFront` for the front panel and `seatBack - waistBack` for the rear panel). This ensures that the total waist width does not change when balancing the waist.

Finally, remove the unnecesary `cfWaist` point from the front panel.